infer.py
"""
infer.py
Synthesizes speech from text using a fine-tuned voice cloning model checkpoint.
"""
import argparse
import torch
from utils.audio import save_audio
# TODO: Add support for Coqui TTS and VITS


def main():
    parser = argparse.ArgumentParser(description="Run voice cloning inference.")
    parser.add_argument('--text', type=str, required=True, help='Text to synthesize')
    parser.add_argument('--voice', type=str, required=True, help='Path to voice checkpoint')
    parser.add_argument('--output', type=str, required=True, help='Output audio file (wav or mp3)')
    args = parser.parse_args()

    # Load checkpoint
    checkpoint = torch.load(args.voice, map_location='cpu')
    # TODO: Load fine-tuned TTS model and synthesize speech
    print("[TODO] Speech synthesis not yet implemented.")
    # Dummy: Save silence
    import numpy as np
    silence = np.zeros(16000)
    save_audio(silence, args.output)
    print(f"Saved dummy audio to {args.output}")

if __name__ == "__main__":
    main()
