requirements.txt
torch
librosa
soundfile
torchaudio
coqpit
TTS
openai-whisper
